pulled out FilterVariants Expr and List functions - moved them to VSM and richVDS - moved generation of Variants RDD from List to Variant
